pyproject.toml
[project]
name = "magpie"
version = "0.1.0"
description = "Local-first CLI + daemon that tags JPEG/HEIC photos with IPTC captions and keywords via vision LLMs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyexiftool>=0.5.6",
    "watchdog>=4.0",
    "typer>=0.12",
    "rich>=13.7",
    "pillow>=10.3",
    "pillow-heif>=0.18",
    "pydantic>=2.7",
    "fastapi>=0.111",
    "uvicorn>=0.30",
]

[project.scripts]
magpie = "magpie.cli:app"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/magpie"]

[tool.hatch.build.targets.wheel.force-include]
"src/magpie/webui/static" = "magpie/webui/static"
